+++ tests/Validator/Constraints/PasswordStrengthValidatorTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Vrok\SymfonyAddons\Tests\Validator\Constraints;
 
 use PHPUnit\Framework\Attributes\DataProvider;
+use Symfony\Component\Validator\Exception\UnexpectedValueException;
 use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 use Vrok\SymfonyAddons\Validator\Constraints\PasswordStrength;
 use Vrok\SymfonyAddons\Validator\Constraints\PasswordStrengthValidator;
@@ -39,9 +40,47 @@ class PasswordStrengthValidatorTest extends ConstraintValidatorTestCase
 
     public function testExpectsStringCompatibleType(): void
     {
-        $this->expectException(\Symfony\Component\Validator\Exception\UnexpectedValueException::class);
+        $this->expectException(UnexpectedValueException::class);
         $constraint = new PasswordStrength();
         $this->validator->validate(new \stdClass(), $constraint);
+    }
+
+    public function testConstraintWithNamedArgument(): void
+    {
+        $constraint = new PasswordStrength(message: 'myMessage');
+
+        $this->validator->validate('fail', $constraint);
+
+        $violation = $this->buildViolation('myMessage')
+            ->setCode(PasswordStrength::PASSWORD_TOO_WEAK_ERROR);
+
+        $violation->assertRaised();
+    }
+
+    public function testConstraintWithMinStrengthOption(): void
+    {
+        $constraint = new PasswordStrength(minStrength: 12);
+
+        // "works" is exactly strength 12
+        $this->validator->validate('works', $constraint);
+
+        $this->assertNoViolation();
+    }
+
+    // @todo remove with SymfonyAddons 3.0
+    public function testConstraintWithOptions(): void
+    {
+        $constraint = new PasswordStrength(['message' => 'myMessage']);
+
+        $this->validator->validate('fail', $constraint);
+
+        $violation = $this->buildViolation('myMessage')
+            ->setCode(PasswordStrength::PASSWORD_TOO_WEAK_ERROR);
+
+        $violation->assertRaised();
+        $this->expectUserDeprecationMessage(
+            'Since symfony/validator 7.3: Passing an array of options to configure the "Vrok\SymfonyAddons\Validator\Constraints\PasswordStrength" constraint is deprecated, use named arguments instead.'
+        );
     }
 
     public function testEmptyStringIsInvalid(): void
@@ -51,6 +90,7 @@ class PasswordStrengthValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate('', $constraint);
 
         $this->buildViolation($constraint->message)
+            ->setCode(PasswordStrength::PASSWORD_TOO_WEAK_ERROR)
             ->assertRaised();
     }
 
@@ -72,6 +112,7 @@ class PasswordStrengthValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($value, $constraint);
 
         $this->buildViolation($constraint->message)
+            ->setCode(PasswordStrength::PASSWORD_TOO_WEAK_ERROR)
             ->assertRaised();
     }
 }